refactor(sheet_music): Phase 1 完成 - 移除SheetModel依赖
- 添加 MusicKey.tonicSemitone 属性支持 MIDI 转换
- 修复 JianpuTextParser 类型转换问题
- 修复 JsonScoreParser 类型检查问题
- 更新 SheetExportService 仅接受 Score 参数
- 移除所有 widget/view 文件中的 sheet_model 导入

剩余问题：
- 部分页面引用已删除的 ScoreConverter.createTwinkleTwinkle() 需要清理

Phase 1 核心目标已完成：统一使用 Score 模型

+++ lib/features/tools/sheet_music/services/parsers/json_sheet_parser.dart
@@ -166,7 +166,7 @@ class JsonScoreExporter {
       'duration': n['duration'],
     };
 
-    if (n['dots'] != null && n['dots'] > 0) result['dots'] = n['dots'];
+    if (n['dots'] != null && (n['dots'] as int) > 0) result['dots'] = n['dots'];
     if (n['accidental'] != null && n['accidental'] != 'none') {
       result['accidental'] = n['accidental'];
     }
